reworking wires

+++ CsHook/mods/mechgame/hooks.c
@@ -36,6 +36,6 @@ void hook_mechgame() {
 
 	
 	// entities
-	CS_npcFuncTable[20] = (int)&NPC_circuit_solver;
-	CS_npcFuncTable[78] = (int)&NPC_circuitSwitch;
+	CS_npcFuncTable[20] = &NPC_circuit_solver;
+	CS_npcFuncTable[78] = &NPC_circuitSwitch;
 }
